Added inline palette module to reorganize BangleEditor
